Replace tests.py
